+++ lib/rubocop/cop/exception_messages_cops.rb
@@ -3,3 +3,4 @@
 require_relative 'exception_messages/message_node'
 require_relative 'exception_messages/casing'
 require_relative 'exception_messages/punctuation'
+require_relative 'exception_messages/redundant_exception_name'